+++ pytsite/flag/__init__.py
@@ -2,7 +2,7 @@
 """
 # Public API
 from . import _widget as widget
-from ._api import count, delete
+from ._api import flag, average, count, delete, is_flagged, sum, toggle, unflag
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -13,12 +13,13 @@ def __init():
     from pytsite import assetman, odm, tpl, lang, events, ajax
     from . import _model, _eh
 
+    # ODM models
     odm.register_model('flag', _model.Flag)
+
+    # Resources
     lang.register_package(__name__)
     tpl.register_package(__name__)
-
     assetman.register_package(__name__)
-    assetman.add('pytsite.flag@js/common.js', permanent=True)
 
     events.listen('pytsite.odm.entity.delete', _eh.odm_entity_delete)