+++ pytsite/flag/_model.py
@@ -13,8 +13,12 @@ class Flag(_odm.Entity):
     def _setup_fields(self):
         """Hook.
         """
+        self.define_field(_odm.field.String('type', default='default'))
         self.define_field(_odm.field.Ref('entity', nonempty=True))
         self.define_field(_odm.field.Ref('author', model='user', nonempty=True))
+        self.define_field(_odm.field.Decimal('score', default=1))
 
     def _setup_indexes(self):
+        """Hook.
+        """
         self.define_index([('entity', _odm.I_ASC), ('author', _odm.I_ASC)], unique=True)