+++ pytsite/odm/_eh.py
@@ -1,0 +1,5 @@
+from pytsite import console as _console
+
+
+def pytsite_update_after():
+    _console.run_command('odm', reindex=True)