+++ pytsite/widget/res/assets/js/traffic-light-score.js
@@ -1,0 +1,3 @@
+$(window).on('pytsite.widget.init:pytsite.widget._select.TrafficLightScore', function (e, widget) {
+    $(window).trigger('pytsite.widget.init:pytsite.widget._select.Score', [widget]);
+});